Refactor project structure to use 'vlax1d' as the main package. Updated import paths in example scripts and tests. Removed unused files and directories, including old egg-info and source files. Added new implementations for equations, plotting, and solver functionalities. Updated setup configuration and .gitignore for better project management.
