+++ examples/two_stream_instability.py
@@ -6,9 +6,9 @@ import matplotlib.pyplot as plt
 import pandas as pd
 from pathlib import Path
 from scipy.stats import linregress
-from src.equations import objective
-from src.solver import solver_ode
-from src.plot import plot_results
+from vlax1d.equations import objective
+from vlax1d.solver import solver_ode
+from vlax1d.plot import plot_results
 
 jax.config.update("jax_enable_x64", True)
