+++ pyproject.toml
@@ -1,4 +1,4 @@
 [build-system]
 requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
-pythonpath = "src"
+pythonpath = "vlax1d"
