+++ setup.py
@@ -2,7 +2,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="VlaX1D",
+    name="vlax1d",
     version="0.1.1",
     author="Jianfeng Ye",
     author_email="xyjjeff23@gmail.com",
@@ -10,8 +10,7 @@ setup(
     long_description=open("README.md", "r", encoding="utf-8").read(),
     long_description_content_type="text/markdown",
     url="https://github.com/XYJeff23/1D_Vlasov_JAX_Solver",
-    packages=find_packages(where="src"),
-    package_dir={"": "src"},
+    packages=find_packages(),
     install_requires=[  # 指定依赖项
         "jax",
         "jaxlib",
