--- src/Oned_Vlasov_Poisson_jax_solver.egg-info/requires.txt
@@ -1,7 +1,0 @@
-jax
-jaxlib
-diffrax
-numpy
-matplotlib
-scipy
-pandas
