+++ vlax1d/equations.py
